fizzbuzzのプログラム本体とテストコードを分離した。
